rename: URLQueryString from urlQueryString
