+++ src/presentation/pages/TimecardReportPage/TimecardReportPage.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { useLoginRedirection } from 'hooks/useLoginRedirection';
 import { WithHeaderLayout } from 'presentation/layouts/WithHeaderLayout/WithHeaderLayout';
 import { getThisMonthDateString } from 'utils/dateUtil';
-import { useUrlQueryString } from 'hooks/useUrlQueryString';
+import { useURLQueryString } from 'hooks/useURLQueryString';
 import { useUserPreference } from 'hooks/useUserPreference';
 import { useDailyTimeRecordsOfMonth } from 'hooks/useDailyTimeRecordsOfMonth';
 import { useMonthlyOverview } from 'hooks/useMonthlyOverview';
@@ -14,7 +14,7 @@ import { Button } from 'presentation/components/inputs/Button/Button';
 import { TextArea } from 'presentation/components/inputs/TextArea/TextArea';
 import { TabItem, Tabs } from 'presentation/components/navigation/Tabs/Tabs';
 import { Head } from 'Head';
-import { createURLQueryParser, getGenericsOrElse, getDateStringOrElse } from 'utils/urlQueryStringUtil';
+import { createURLQueryParser, getGenericsOrElse, getDateStringOrElse } from 'utils/URLQueryStringUtil';
 import { DATE_FORMAT } from 'constants/dateFormat';
 
 const ReportTypes = ['weeklyReport', 'monthlyReport'] as const;
@@ -44,7 +44,7 @@ const TimecardReportPage = () => {
 
   const { uid } = useParams<{ uid: string }>();
 
-  const [{ month: selectedMonth, type: reportType }, setSearchParams] = useUrlQueryString({
+  const [{ month: selectedMonth, type: reportType }, setSearchParams] = useURLQueryString({
     parser,
   });
